added route refactor
